fix warning
Former-commit-id: 9ae7e3c40c4a53f9cf1a64d61b2408df0c40f30d

+++ src/policy/feerate.cpp
@@ -22,7 +22,7 @@ CAmount CFeeRate::GetFee(size_t nBytes_) const
 {
     if((nBytes_*1000) > uint64_t(std::numeric_limits<int64_t>::max()))
         return 0;
-    if((nSatoshisPerK*1000) > uint64_t(std::numeric_limits<int64_t>::max()))
+    if((nSatoshisPerK*1000) > std::numeric_limits<int64_t>::max())
         return 0;
     assert(nBytes_ <= uint64_t(std::numeric_limits<int64_t>::max()));
     int64_t nSize = int64_t(nBytes_);
